Adding deadgrep and updating wgrep
I often use `consult-rg` (and will continue to do so).  My use case for
`consult-rg` is the live search paired with result preview.

Deadgrep sits closer to a transient menu, in which I write my search
result and then have a buffer to continue to refine.

I can see both of these serving a purpose.