fix(): favor eglot built-in
I was encountering an error in which the straight version of eglot was
ahead of the built-in; and was breaking functionality with
eldoc/flymake.  Namely I was getting the following error:

```
eldoc error: (wrong-type-argument characterp compiler)
```

I did some triage, stripping my init.el down to:

```
(require 'go-ts-mode)
(use-package eglot :load-path "~/.emacs.d/straight/build/eglot")
```

That failed in the same manner.  When I replaced the `use-package` with
`require 'eglot` everything worked.